requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.23.1
python-multipart==0.0.6
#transformers==4.41.0
langchain==1.0.0
#faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==0.8.11
aiofiles==23.1.0
chromadb==1.0.20
anyio==3.7.1
numpy==2.1.0
#langsmith==0.0.52
pydantic==2.11.5
requests==2.32.5
python-dotenv==1.0.1
openai==1.109.1
langchain-chroma==1.0.0
langchain-text-splitters==1.0.0
langchain-community==0.4.1
langchain-openai==1.0.0
#langchain-google-genai==1.0.10
